% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcdosestat.R
\name{calcdosestat}
\alias{calcdosestat}
\title{Statistical analysis for metabolomic dose-response analysis (TOXcms step 1)}
\usage{
calcdosestat(Feature, Dose_Levels, multicomp = c("none","ttest","tukey","games-howell"),
p.adjust.method = c("none","holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr"),projectName = "dose_response_analysis",...)
}
\arguments{
\item{Feature}{data.table a feature table including basic information (m/z, retention time, etc) and MS raw intensities among all dose levels.}

\item{Dose_Levels}{character Keywords indicating ordered dose levels in feature table. The dose levels should be ordered incrementally.}

\item{multicomp}{options for multigroup testing of significant difference. "none"- Welch t-test on adjacent doses; "ttest" - pairwise Welch t-test among all doses;
"tukey"/"games-howell" - post-hoc tests followed by ANOVA.}

\item{p.adjust.method}{p-value adjustment for false-positive reduction in multiple testing. See also \link[stats]{p.adjust}}

\item{projectName}{character Project name of the TOXcms analysis}

\item{...}{Further arguments to be passed to calcdosestat.}
}
\value{
calcdosestat returns a list object consisting all the statistical results described above.
}
\description{
The function calcdosestat() performs statistical analysis on metabolic responses at diffrent dose levels. Basic statistics of each dose level is calculated.
MS raw intensities of each feature are normalized by range scaling (x-min/max-min). ANOVA and multigroup testing are performed to obtain p-values and relative changes
among dose levels. The output is called dosestat.
}
\author{
Cong-Hui Yao <conghui.yao@wustl.edu>
Lingjue Wang (Mike) <wang.lingjue@wustl.edu>
}
