% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr4pl_fit.R
\name{dr4pl_fit}
\alias{dr4pl_fit}
\title{Fitting dose response four parameter logistic function to dose-dependent data points}
\usage{
dr4pl_fit(DoseResponse_report, ED=0.5, Dose_values, export = TRUE)
}
\arguments{
\item{DoseResponse_report}{list a complex list of dose-dependent features processed by trendfilter function.}

\item{ED}{numeric effective dose value from 0 to 1. 0.5 by default.}

\item{Dose_values}{dose values to be fitted.}

\item{export}{TRUE/FALSE whether to export the fitting results to a CSV file.}
}
\description{
This function applies dr4pl function to fitting a dose-response four parameter logistic model to 
each dose-dependent metabolomic features.
}
