% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcdosestat.R
\name{normalization}
\alias{normalization}
\title{Row-wise data normalization and transformation}
\usage{
normalization(data,Norm.method=c("range","auto","pareto","vast","level",
"log10","log2","sqrt"),...)
}
\arguments{
\item{data}{a matrix of numerical values.}

\item{Norm.method}{normalization methods to be applied. By default using range scaling. See details for further information.}

\item{...}{Further arguments to be passed to normalization.}
}
\description{
Perform row-wise data normalization or transformation of a given data.matrix. Return a data.matrix with normalized values.
}
\details{
normalization() applies a series of commonly used metrices for data normalization. Range scaling ("range") focuses on data correlation and restricted the values in between 0 to 1, (x-min)/(max-min);
Auto scaling ("auto") focuses on data correlation and data is centered to 0 with a standard deviation of 1, (x-mean)/std; Pareto scaling ("pareto") focuses on data correlation and discriminates large fold-chanegs, (x-mean)/(sqrt(std));
Vast scaling ("vast") focuses on less varying data and discriminates largely varying data, (x-mean)/(std*cv); Level scaling ("level") focuses on fold changes against the mean value, (x-mean)/mean;
log10 or log2 transformation ("log10","log2") focuses on scaling the exponential relationship to a linear model, and centering data at the mean; squart root ("sqrt") is a pesudo scaling for positive values only.
}
