% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{ObservedED}
\alias{ObservedED}
\title{Calculate observed ED values using dr4pl-fitted parameters.}
\usage{
ObservedED(ED, theta)
}
\arguments{
\item{ED}{numeric the level of effective dose from 0 to 1.}

\item{theta}{the parameter set of logistic regression curve.}
}
\description{
Calculate observed ED values using dr4pl-fitted parameters.
}
