% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotpca.R
\name{plotpca}
\alias{plotpca}
\title{Principle component analysis of all features color-coded by ED50 value prediction}
\usage{
plotpca(DoseResponse_report, DoseStat, EDrange = c(0,200),...)
}
\arguments{
\item{DoseResponse_report}{list The output object of fitdrc() including ED50 values. See \code{\link{calcdosestat}}}

\item{DoseStat}{list The output object of calcdosestat(). See \code{\link{fitdrc}}}

\item{EDrange}{numeric The range of minimum and maximum ED50 value to be shown in the PCA plots.}

\item{...}{Further arguments to be passed to plotpca.}
}
\value{
plotpca function does not return any objects but generates a PDF with principal component plots of non-significant features (grey) and dose-response features with color-coded ED50 values.
}
\description{
plotpca function generates a PCA plot showing all input features and highlighting dose-dependent features with color-coded ED50 values.
}
\details{
plotPCA takes the raw intensity values and calculates the mean intensity for each feature for each dose. After this is calculated
range scaling is applied to make all intensities between zero and one. With the normalized mean intensities PCA is performed and the first
two components are plotted. Each dot represents a feature with the color representing the ED50 value. Grey dots were not determined to be
following a dose reponse trend.
}
\author{
Ethan Stancliff Yao <estancliff@wustl.edu>
Lingjue Wang (Mike) <wang.lingjue@wustl.edu>
}
