% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottrend.R
\name{plottrend}
\alias{plottrend}
\title{A supplemental function to visualize all trends (TOXcms step 3)}
\usage{
plottrend(DoseResponse_report, Dose_conditions=NULL, y_transform=TRUE, mz_tag = "mzmed", rt_tag = "rtmed",file="metablictrend.pdf")
}
\arguments{
\item{DoseResponse_report}{list The output list object from trendfilter().}

\item{Dose_conditions}{character Labels at the x axis in the metabolic trend plot, indicating the conditions.}

\item{y_transform}{TRUE/FALSE Whether log2 transformation should be applied to scale y axis.}

\item{mz_tag}{character Name of the m/z column in the feature table.}

\item{rt_tag}{character Name of the retention time column in the feature table.}

\item{file}{character The filename of the exported PDF.}
}
\value{
The function is a plotting function and dose not have a return.
}
\description{
The plottrend() plots monotonic and reverse trends based on results from trendfiler(): monotonic ("increase","decrease","mono") trend and reverse trend.
The function export a PDF file including all the plots. plottrend() does not have a return .
}
\author{
Cong-Hui Yao <conghui.yao@wustl.edu>
Lingjue Wang (Mike) <wang.lingjue@wustl.edu>
}
