% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxcms.r
\docType{package}
\name{toxcms-package}
\alias{toxcms}
\alias{toxcms-package}
\title{toxcms: An Informatic Tool to Perform Dose-response Metabolomic Analysis}
\description{
An informatic tool for metabolomics data analysis. TOXcms profiles metabolic trends as a function of drug doses from untargeted metabolomics data, models and visualizes dose-dependent metabolic changes. TOXcms is useful in screening metabolic targets of drugs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pattilab/toxcms}
  \item Report bugs at \url{https://github.com/pattilab/toxcms/issues}
}

}
\author{
\strong{Maintainer}: Lingjue (Mike) Wang \email{wang.lingjue@wustl.edu}

Authors:
\itemize{
  \item Cong-Hui Yao \email{conghui.yao@wustl.edu}
}

Other contributors:
\itemize{
  \item Ethan Stancliffe \email{estancliffe@wustl.edu} [contributor]
}

}
\keyword{internal}
