% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusttrend.R
\name{clusttrend}
\alias{clusttrend}
\title{Clustering trend for metabolomic dose-response analysis (TOXcms step 3)}
\usage{
clusttrend(DoseResponse_report, reference_index = NULL,
  sort.method = c("clust", "layer"), sort.thres = 20,
  dist.method = "euclidean", hclust.method = "average", mztag = "mzmed",
  rttag = "rtmed", heatmap.on = FALSE, plot.all = FALSE, export = TRUE,
  filename = "Heatmap.pdf", ...)
}
\arguments{
\item{DoseResponse_report}{This is the output of trendfilter(), the end of toxcms step 2.}

\item{reference_index}{This is the indices of the reference trends. clusttrend() apply the algorithm for each of the reference index.}

\item{sort.method}{a two-element vector indicating the sorting methods. The first element can be either "dist" or "clust". If the first element is "dist", the second element can be "far","near" or "both".
If the first element is "clust", then the second-element can be “range" or "layer".}

\item{sort.thres}{a numeric number, indicating the sorting threshold for each group. Can be a fraction number smaller than 1 or a positibe integer larger than 1.}

\item{dist.method}{Matrice used for similarity calculation. See also \link[stats]{dist}.}

\item{hclust.method}{Methods for hierarchical clustering. See also \link[stats]{hlust}/}

\item{mztag}{name of the m/z column in the feature table.}

\item{rttag}{name of the retention time column in the feature table.}

\item{heatmap.on}{a TRUE/FALSE value indicating whether heatmap shoudld be generated.}

\item{plot.all}{a TRUE/FALSE value indicating whether a heatmap of all trends should be generated.}

\item{export}{a TRUE/FALSE value indicating whetehr the trend clusters should be exported to a csv. file.}

\item{filename}{a character indicating the name of the exported csv file.}
}
\value{
clusttrend returns a list object, within wich each element is a trend cluster. The name of each element is the reference_index. Heatmap visualization of all trend clusters will be generated if heatmap.on=TRUE.
}
\description{
This function applies similarity matrices and hierarchical clustering to group metabolomic trends based on the similarity and dissimilarity in their dose-dependent responses.
}
\details{
clusttrend uses either distance similarity for
}
\examples{
#Extract 20 most similar trends of trend #308, #61 and #85 by hierarchical clustering on trend. Distance matrice and Clustering method are default.
clust <- clusttrend(doseresponse_report, reference_index = c(308,61,85), sort.method = c("clust","range")), sort.thres = 20, mztag ="mzmed", rttag="rtmed". heatmap.on=TRUE, plot.all=TRUE, export=TRUE, filename="trendcluster_20.pdf")
#Extract 10\% of all trends that are most similar to trend #308, #61, #85 by manhattan distance matrice.
clust <- clusttrend(doseresponse_report, reference_index = c(308,61,85), sort.method = c("dist","range")), sort.thres = 20, mztag ="mzmed", rttag="rtmed". heatmap.on=TRUE, plot.all=TRUE, export=TRUE, filename="trendcluster_20.pdf")
}
\author{
Lingjue Wang (Mike) <wang.lingjue@wustl.edu>
Cong-Hui Yao <conghui.yao@wustl.edu>
}
