% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendfilter.R
\name{trendfilter}
\alias{trendfilter}
\title{Trend filtering for metabolomic dose-response analysis (TOXcms step 2)}
\usage{
trendfilter(DoseStat, pval_cutoff = 0.05, pval_thres = 1, anova_cutoff = 0.05,trend =c("increase","decrease","mono","reverse","all"),
relChange_cutoff = NULL, export = FALSE)
}
\arguments{
\item{DoseStat}{the output list object of calcdosesstat().}

\item{pval_cutoff}{a filtering threshold of significance cutoff of Welch t-test or post-hoc tests.}

\item{pval_thres}{a filtering threshold of numbers of significant dose pairs. With n dose levels, maximum value for adjacent comparison (multicomp="none") is n-1, maximum value for multi-group comparison
(multicomp="ttest"/"tukey"/"games-howell") is n(n-1)/2.}

\item{anova_cutoff}{a filtering threshold of significancce cutoff of ANOVA.}

\item{trend}{Type of trend. Options are "increase", "decrease", "mono" (both increase/decrease), "reverse" (one inflection point, "V" or "∧" shape), "all" (no filter).}

\item{relChange_cutoff}{a filtering threshold for monotonic trend ("increase"/"decrease"/"mono"). This filter removes the trends with non-significant relative change larger than the threshold and opposed to desired trends.}

\item{export}{a TRUE/FALSE value. When set to TRUE, a csv file including all results will be exported.}
}
\value{
The function returns doseresponse_report, a list object including same elements as the input that are filtered based on trend and statistical thresholds.
}
\description{
The trendfilter() filters two pre-defined trends (dose-response models) from data processed by calcdosestat: monotonic ("increase","decrease","mono") trend and reverse trend.
The function returns the same list object as the input with only filtered trends retain. The resulting object is called doseResponse_report.
}
\author{
Cong-Hui Yao <conghui.yao@wustl.edu>
Lingjue Wang (Mike) <wang.lingjue@wustl.edu>
}
